docker to aws

+++ apps/core_microservice/src/migrations/1767610953846-add-bio-and-avatar-to-users.ts
@@ -4,10 +4,18 @@ export class AddBioAndAvatarToUsers1767610953846 implements MigrationInterface {
   name = 'AddBioAndAvatarToUsers1767610953846';
 
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query('ALTER TABLE "users" DROP COLUMN "website"');
-    await queryRunner.query('ALTER TABLE "users" DROP COLUMN "avatar"');
-    await queryRunner.query('ALTER TABLE "users" ADD "avatar_url" character varying(255)');
-    await queryRunner.query('ALTER TABLE "users" DROP COLUMN "bio"');
+    if (await queryRunner.hasColumn('users', 'website')) {
+      await queryRunner.query('ALTER TABLE "users" DROP COLUMN "website"');
+    }
+    if (await queryRunner.hasColumn('users', 'avatar')) {
+      await queryRunner.query('ALTER TABLE "users" DROP COLUMN "avatar"');
+    }
+    if (!(await queryRunner.hasColumn('users', 'avatar_url'))) {
+      await queryRunner.query('ALTER TABLE "users" ADD "avatar_url" character varying(255)');
+    }
+    if (await queryRunner.hasColumn('users', 'bio')) {
+      await queryRunner.query('ALTER TABLE "users" DROP COLUMN "bio"');
+    }
     await queryRunner.query('ALTER TABLE "users" ADD "bio" text');
   }
